Design: search UI
1
