+++ app/src/main/java/com/example/mumuk/ui/category/CategoryItem.kt
@@ -1,3 +1,3 @@
-package com.example.mumuk.Category
+package com.example.mumuk.ui.category
 
 data class CategoryItem(val name: String, val imageResId: Int)
