+++ app/src/main/java/com/example/mumuk/ui/category/CategoryRecipeCard.kt
@@ -1,4 +1,4 @@
-package com.example.mumuk.Category
+package com.example.mumuk.ui.category
 
 data class CategoryRecipeCard(
     val label: String,
