+++ app/src/main/java/com/example/mumuk/ui/category/CategoryRecipeCardAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.mumuk.Category
+package com.example.mumuk.ui.category
 
 import android.view.LayoutInflater
 import android.view.View
