+++ app/src/main/java/com/example/mumuk/ui/category/Health/CategoryHealthFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mumuk.Category.Weight
+package com.example.mumuk.ui.category.Health
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,18 +7,18 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.mumuk.Category.CategoryFragment
-import com.example.mumuk.Category.CategoryItem
-import com.example.mumuk.Category.CategoryItemAdapter
+import com.example.mumuk.ui.category.CategoryFragment
+import com.example.mumuk.ui.category.CategoryItem
+import com.example.mumuk.ui.category.CategoryItemAdapter
 import com.example.mumuk.R
 
-class CategoryWeightFragment : Fragment() {
+class CategoryHealthFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_category_weight, container, false)
+        return inflater.inflate(R.layout.fragment_category_health, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -30,15 +30,17 @@ class CategoryWeightFragment : Fragment() {
         recyclerView.adapter = CategoryItemAdapter(getDummyList()) { item ->
             val selectedTab = item.name
             (parentFragment as? CategoryFragment)?.showFullScreenFragment(
-                CategoryWeightLossFragment.newInstance(selectedTab)
+                CategoryLowSugarFragment.newInstance(selectedTab)
             )
         }
     }
 
     private fun getDummyList(): List<CategoryItem> {
         return listOf(
-            CategoryItem("체중 감량", R.drawable.ic_category_weight_1),
-            CategoryItem("근육 증가", R.drawable.ic_category_weight_2)
+            CategoryItem("당 줄이기", R.drawable.ic_category_health_1),
+            CategoryItem("혈압관리", R.drawable.ic_category_health_2),
+            CategoryItem("콜레스테롤 관리", R.drawable.ic_category_health_3),
+            CategoryItem("소화 건강", R.drawable.ic_category_health_4)
         )
     }
 }
