+++ app/src/main/java/com/example/mumuk/ui/category/Health/CategoryLowSugarFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mumuk.Category.Random
+package com.example.mumuk.ui.category.Health
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,20 +8,20 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.mumuk.Category.CategoryRecipeCard
-import com.example.mumuk.Category.CategoryRecipeCardAdapter
+import com.example.mumuk.ui.category.CategoryRecipeCard
+import com.example.mumuk.ui.category.CategoryRecipeCardAdapter
 import com.example.mumuk.R
 import com.google.android.material.tabs.TabLayout
 
-class CategoryRandomRecipeFragment : Fragment() {
+class CategoryLowSugarFragment : Fragment() {
 
     private lateinit var recyclerView: RecyclerView
     private lateinit var tabLayout: TabLayout
     private var selectedTabTitle: String? = null
 
     companion object {
-        fun newInstance(selectedTab: String): CategoryRandomRecipeFragment {
-            val fragment = CategoryRandomRecipeFragment()
+        fun newInstance(selectedTab: String): CategoryLowSugarFragment {
+            val fragment = CategoryLowSugarFragment()
             val args = Bundle()
             args.putString("selected_tab", selectedTab)
             fragment.arguments = args
@@ -36,7 +36,7 @@ class CategoryRandomRecipeFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_category_random_recipe, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_category_low_sugar, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -49,17 +49,23 @@ class CategoryRandomRecipeFragment : Fragment() {
         recyclerView = view.findViewById(R.id.category_recipe_recycler_view)
         recyclerView.layoutManager = GridLayoutManager(context, 2)
 
-        setupSingleTab()
-        setupTabListener()
+        setupCustomTabs()
+        setupTabSelectionListener()
     }
 
-    private fun setupSingleTab() {
-        val tabName = selectedTabTitle ?: "랜덤식단"
-        val tab = tabLayout.newTab()
-        tabLayout.addTab(tab)
-        tab.customView = createCustomTabView(tabName, true)
-        tab.select()
-        updateRecyclerWith(tabName)
+    private fun setupCustomTabs() {
+        val tabs = listOf("당 줄이기", "혈압관리", "콜레스테롤 관리", "소화 건강")
+        val initialTab = selectedTabTitle ?: "당 줄이기"
+
+        for (title in tabs) {
+            val tab = tabLayout.newTab()
+            tabLayout.addTab(tab)
+            tab.customView = createCustomTabView(title, selected = (title == initialTab))
+            if (title == initialTab) {
+                tab.select()
+                updateRecyclerWith(title)
+            }
+        }
     }
 
     private fun createCustomTabView(title: String, selected: Boolean): View {
@@ -70,11 +76,12 @@ class CategoryRandomRecipeFragment : Fragment() {
         return view
     }
 
-    private fun setupTabListener() {
+    private fun setupTabSelectionListener() {
         tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
             override fun onTabSelected(tab: TabLayout.Tab?) {
-                tab?.customView?.findViewById<TextView>(R.id.tab_text)?.isSelected = true
-                updateRecyclerWith(tab?.customView?.findViewById<TextView>(R.id.tab_text)?.text.toString())
+                val textView = tab?.customView?.findViewById<TextView>(R.id.tab_text)
+                textView?.isSelected = true
+                updateRecyclerWith(textView?.text.toString())
             }
 
             override fun onTabUnselected(tab: TabLayout.Tab?) {
@@ -87,10 +94,21 @@ class CategoryRandomRecipeFragment : Fragment() {
 
     private fun updateRecyclerWith(tabName: String) {
         val items = when (tabName) {
-            "랜덤식단" -> listOf(
-                CategoryRecipeCard("랜덤식단", "연어 포케"),
-                CategoryRecipeCard("랜덤식단", "연어 포케"),
-                CategoryRecipeCard("랜덤식단", "연어 포케")
+            "당 줄이기" -> listOf(
+                CategoryRecipeCard("당줄이기", "연어 포케"),
+                CategoryRecipeCard("당줄이기", "연어 포케")
+            )
+            "혈압관리" -> listOf(
+                CategoryRecipeCard("혈압관리", "연어 포케"),
+                CategoryRecipeCard("혈압관리", "연어 포케")
+            )
+            "콜레스테롤 관리" -> listOf(
+                CategoryRecipeCard("콜레스테롤", "연어 포케"),
+                CategoryRecipeCard("콜레스테롤", "연어 포케")
+            )
+            "소화 건강" -> listOf(
+                CategoryRecipeCard("소화건강", "연어 포케"),
+                CategoryRecipeCard("소화건강", "연어 포케")
             )
             else -> emptyList()
         }
