+++ app/src/main/java/com/example/mumuk/ui/category/Weight/CategoryWeightFragment.kt
@@ -1,4 +1,4 @@
-package com.example.mumuk.Category.Random
+package com.example.mumuk.ui.category.Weight
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,18 +7,18 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.mumuk.Category.CategoryFragment
-import com.example.mumuk.Category.CategoryItem
-import com.example.mumuk.Category.CategoryItemAdapter
+import com.example.mumuk.ui.category.CategoryFragment
+import com.example.mumuk.ui.category.CategoryItem
+import com.example.mumuk.ui.category.CategoryItemAdapter
 import com.example.mumuk.R
 
-class CategoryRandomFragment : Fragment() {
+class CategoryWeightFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_category_random, container, false)
+        return inflater.inflate(R.layout.fragment_category_weight, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -30,14 +30,15 @@ class CategoryRandomFragment : Fragment() {
         recyclerView.adapter = CategoryItemAdapter(getDummyList()) { item ->
             val selectedTab = item.name
             (parentFragment as? CategoryFragment)?.showFullScreenFragment(
-                CategoryRandomRecipeFragment.newInstance(selectedTab)
+                CategoryWeightLossFragment.newInstance(selectedTab)
             )
         }
     }
 
     private fun getDummyList(): List<CategoryItem> {
         return listOf(
-            CategoryItem("랜덤식단", R.drawable.ic_category_random_1)
+            CategoryItem("체중 감량", R.drawable.ic_category_weight_1),
+            CategoryItem("근육 증가", R.drawable.ic_category_weight_2)
         )
     }
 }
