+++ app/src/main/res/drawable/ic_search.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="26dp"
-    android:height="25dp"
-    android:viewportWidth="26"
-    android:viewportHeight="25">
+    android:width="21dp"
+    android:height="20dp"
+    android:viewportWidth="21"
+    android:viewportHeight="20">
   <path
-      android:pathData="M19.232,16.123H18.073L17.663,15.727C18.579,14.663 19.249,13.41 19.624,12.057C19.999,10.704 20.07,9.285 19.833,7.901C19.144,3.827 15.742,0.574 11.636,0.075C10.192,-0.107 8.726,0.043 7.35,0.514C5.973,0.985 4.723,1.764 3.694,2.792C2.665,3.82 1.885,5.07 1.414,6.446C0.943,7.822 0.793,9.287 0.975,10.73C1.474,14.833 4.729,18.233 8.806,18.922C10.19,19.159 11.61,19.088 12.964,18.713C14.318,18.338 15.571,17.669 16.636,16.753L17.032,17.163V18.321L23.264,24.549C23.865,25.15 24.848,25.15 25.449,24.549C26.05,23.948 26.05,22.967 25.449,22.366L19.232,16.123ZM10.433,16.123C6.782,16.123 3.835,13.177 3.835,9.528C3.835,5.879 6.782,2.933 10.433,2.933C14.085,2.933 17.032,5.879 17.032,9.528C17.032,13.177 14.085,16.123 10.433,16.123Z"
-      android:fillColor="#000000"/>
+      android:pathData="M15.165,12.898H14.239L13.91,12.582C14.643,11.73 15.179,10.728 15.479,9.645C15.779,8.563 15.836,7.428 15.646,6.321C15.095,3.062 12.373,0.459 9.089,0.06C7.934,-0.086 6.761,0.034 5.66,0.411C4.559,0.788 3.558,1.411 2.735,2.234C1.912,3.056 1.288,4.056 0.911,5.157C0.534,6.257 0.414,7.43 0.56,8.584C0.959,11.866 3.563,14.586 6.825,15.137C7.932,15.327 9.068,15.27 10.151,14.97C11.234,14.67 12.237,14.135 13.089,13.402L13.406,13.731V14.657L18.391,19.639C18.872,20.12 19.658,20.12 20.139,19.639C20.62,19.159 20.62,18.373 20.139,17.893L15.165,12.898ZM8.127,12.898C5.206,12.898 2.848,10.542 2.848,7.622C2.848,4.703 5.206,2.346 8.127,2.346C11.048,2.346 13.406,4.703 13.406,7.622C13.406,10.542 11.048,12.898 8.127,12.898Z"
+      android:fillColor="#A2A2A2"/>
 </vector>
