+++ app/src/main/res/layout/fragment_category_random_recipe.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".Category.Random.CategoryRandomRecipeFragment">
+    tools:context=".ui.category.Random.CategoryRandomRecipeFragment">
 
     <ImageView
         android:id="@+id/category_back_btn"
