+++ app/src/main/res/layout/fragment_search.xml
@@ -3,12 +3,95 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".ui.home.HomeFragment"
     android:background="@color/white">
 
+    <View
+        android:id="@+id/search_edit_frame"
+        android:layout_width="370dp"
+        android:layout_height="32dp"
+        android:background="@drawable/bg_search_frame"
+        tools:layout_editor_absoluteX="16dp"
+        tools:layout_editor_absoluteY="16dp" />
+
+    <EditText
+        android:id="@+id/sign_up_id_et"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:background="@null"
+        android:hint="검색어를 입력하세요"
+        android:inputType="text"
+        android:textSize="16sp"
+        android:layout_marginStart="16dp"
+        app:layout_constraintStart_toStartOf="@id/search_edit_frame"
+        app:layout_constraintBottom_toBottomOf="@+id/search_edit_frame"
+        app:layout_constraintTop_toTopOf="@id/search_edit_frame"/>
+    
+    <ImageView
+        android:id="@+id/search_btn"
+        android:layout_width="20dp"
+        android:layout_height="20dp"
+        android:src="@drawable/ic_search"
+        android:layout_marginEnd="10dp"
+        app:layout_constraintEnd_toEndOf="@id/search_edit_frame"
+        app:layout_constraintTop_toTopOf="@id/search_edit_frame"
+        app:layout_constraintBottom_toBottomOf="@id/search_edit_frame"/>
+
     <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:text="Search Fragment" />
+        android:id="@+id/search_recently_tv"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="최근 검색어"
+        android:textStyle="bold"
+        android:textSize="20sp"
+        android:layout_marginStart="5dp"
+        android:layout_marginTop="20dp"
+        app:layout_constraintTop_toBottomOf="@id/search_edit_frame"
+        app:layout_constraintStart_toStartOf="@id/search_edit_frame" />
+
+    <FrameLayout
+        android:id="@+id/search_recently_container_fl"
+        android:layout_width="90dp"
+        android:layout_height="35dp"
+        android:layout_marginTop="16dp"
+        android:background="@drawable/bg_search_frame"
+        app:layout_constraintStart_toStartOf="@id/search_recently_tv"
+        app:layout_constraintTop_toBottomOf="@id/search_recently_tv" >
+
+        <TextView
+            android:id="@+id/search_recently_container_text_tv"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="포케"
+            android:textSize="20sp"
+            android:textColor="@color/darker_gray"
+            android:layout_marginStart="8dp"
+            android:layout_marginBottom="3dp"
+            app:layout_constraintStart_toStartOf="@id/search_recently_container_fl"
+            app:layout_constraintTop_toTopOf="@id/search_recently_container_fl"
+            app:layout_constraintBottom_toBottomOf="@id/search_recently_container_fl"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/ic_search_delete"
+            android:layout_marginEnd="8dp"
+            app:layout_constraintEnd_toEndOf="@id/search_recently_container_fl"
+            app:layout_constraintTop_toTopOf="@id/search_recently_container_fl"
+            app:layout_constraintBottom_toBottomOf="@id/search_recently_container_fl"/>
+    
+    <TextView
+        android:id="@+id/search_recommendation_tv"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="추천 검색어"
+        android:textStyle="bold"
+        android:textSize="20sp"
+        android:layout_marginStart="5dp"
+        android:layout_marginTop="20dp"
+        app:layout_constraintTop_toBottomOf="@id/search_recently_container_fl"
+        app:layout_constraintStart_toStartOf="@id/search_recently_container_fl" />
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
