+++ app/src/main/res/values/colors.xml
@@ -2,6 +2,9 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="red">#FF0000</color>
+    <color name="gray">#E0E0E0</color>
+    <color name="darker_gray">#9E9E9E</color>
     <color name="mainColor">#007251</color>
     <color name="category_tab_selected_bg">#FFA552</color>
     <color name="category_tab_unselected_bg">#CCCCCC</color>
